code/flask/flask-docker/.history/requirements_20230522182120.txt
amqp==5.1.1
APScheduler==3.9.1
billiard==3.6.4.0
black==22.8.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Flask==2.2.2
Flask-Cors==3.0.10
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
kombu==5.2.4
MarkupSafe==2.1.1
mypy-extensions==0.4.3
pathspec==0.10.1
platformdirs==2.5.2
prompt-toolkit==3.0.31
pytz==2022.2.1
six==1.16.0
tomli==2.0.1
typing_extensions==4.3.0
vine==5.0.0
waitress==2.1.2
wcwidth==0.2.5
Werkzeug==2.2.2
zipp==3.8.1
gevent==22.10.2
pyOpenSSL==22.1.0
flasgger==0.9.5
marshmallow==3.19.0
marshmallow-dataclass==8.5.14
PyJWT==1.7.1

